accessors refactor
