+++ package.json
@@ -1,14 +1,9 @@
 {
-    "name": "niagads-viz-react-js",
+    "name": "niagads-viz-js",
     "version": "0.1.0",
-    "main": "index.js",
-    "module": "index.js",
     "author": "fossilfriend,bug-sam",
     "scripts": {
-        "dev": "NODE_OPTIONS='--inspect' next dev",
-        "build": "next build",
-        "start": "next start",
-        "lint": "next lint"
+        "build": "tsc"
     },
     "devDependencies": {
         "@types/highcharts": "^7.0.0",
@@ -57,6 +52,7 @@
         "match-sorter": "^6.3.1",
         "next": "^14.1.0",
         "react-csv": "^2.2.2",
-        "react-router-dom": "^6.19.0"
+        "react-router-dom": "^6.19.0",
+        "rollup": "^4.12.0"
     }
 }
