+++ src/components/Chart/Highcharts/HighchartsTrellisPlot.scss
