+++ src/components/Chart/Highcharts/PairedHighchartPlot.tsx
