+++ src/components/Chart/LocusZoom/CustomLocusZoomAdapters.tsx
