+++ src/components/Chart/LocusZoom/index.ts
