+++ src/components/Table/index.ts
@@ -1,0 +1,4 @@
+export * from "./utils";
+export * from "./types"
+export * from "./ColumnAccessors";
+export { default as Table } from "./Table";
