+++ src/components/TextRenderer/Link.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { ColumnAccessor } from "../common/types";
+import { TextRenderer } from "@renderers/types";
 
 export const resolveLink = (url: string, value: string) => {
     return (
@@ -10,9 +10,9 @@ export const resolveLink = (url: string, value: string) => {
 };
 
 // text with tooltip value = { value: string, url: string, tooltip: string}
-export const LinkAccessor: React.FC<ColumnAccessor> = ({ value }) => {
+export const Link: React.FC<TextRenderer> = ({ value }) => {
     return Array.isArray(value) ? (
-        <LinkListAccessor value={value} />
+        <LinkList value={value} />
     ) : "tooltip" in value && value.tooltip != "" ? (
         <div title={value.tooltip} arial-label={value.tooltip}>
             {resolveLink(value.url, value.value)}
@@ -24,11 +24,11 @@ export const LinkAccessor: React.FC<ColumnAccessor> = ({ value }) => {
 
 // json array of [{url: , value: , tooltip?: }, ...]
 // asString returns it as a " // " separated list
-export const LinkListAccessor: React.FC<ColumnAccessor> = ({ value }) => {
+export const LinkList: React.FC<TextRenderer> = ({ value }) => {
     return value.map((item: any, i: number) => (
         <span key={i}>
             {i > 0 && " // "}
-            <LinkAccessor value={item} />
+            <Link value={item} />
         </span>
     ));
 };
