+++ src/components/TextRenderer/NASpan.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+import { TextRenderer } from "@renderers/types";
+
+export const NASpan: React.FC<TextRenderer> = ({value="N/A", className="grey"}) => {
+    return (
+        <span className={className}>
+            {value}
+        </span>
+    );
+};
+
