+++ src/components/index.ts
@@ -1,0 +1,1 @@
+export { Table } from "@table/index"
