--- src/components/table/TableAccessors/index.ts
@@ -1,8 +1,0 @@
-export * from "./ColumnAccessorUtils";
-export * from "./NASpan";
-export * from "./TextAccessors";
-export * from "./SparkPercentageBarAccessor";
-export * from "./BooleanCheckAccessor";
-export * from "./JSONAccessor";
-export * from "./LinkAccessor";
-export * from "./RowSelectButtonAccessor";
